+++ src/main/java/subway/domain/line/Lines.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import subway.exception.DuplicatedLineNameException;
 import subway.exception.StationNotFoundException;
