+++ src/main/java/subway/domain/line/edge/StationEdge.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line.edge;
 
 import subway.exception.IllegalStationEdgeException;
