+++ src/main/java/subway/domain/station/Stations.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.station;
 
 import subway.exception.DuplicatedStationNameException;
 import subway.exception.StationNotFoundException;
