+++ src/main/java/subway/repository/LineRepository.java
@@ -1,6 +1,6 @@
 package subway.repository;
 
-import subway.domain.Line;
+import subway.domain.line.Line;
 
 import java.util.List;
 import java.util.Optional;
