+++ src/main/java/subway/repository/StationRepository.java
@@ -2,7 +2,7 @@ package subway.repository;
 
 import java.util.List;
 import java.util.Optional;
-import subway.domain.Station;
+import subway.domain.station.Station;
 
 public interface StationRepository {
