+++ src/main/java/subway/service/LineService.java
@@ -2,9 +2,9 @@ package subway.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.domain.Line;
+import subway.domain.line.Line;
 import subway.domain.LineDirection;
-import subway.domain.Station;
+import subway.domain.station.Station;
 import subway.domain.Subway;
 import subway.exception.LineNotFoundException;
 import subway.repository.LineRepository;
