+++ src/main/java/subway/ui/LineController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.domain.Line;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.station.Station;
 import subway.ui.dto.LineRequest;
 import subway.ui.dto.LineResponse;
 import subway.ui.dto.StationInsertRequest;
