+++ src/main/java/subway/ui/dto/LineResponse.java
@@ -1,7 +1,7 @@
 package subway.ui.dto;
 
-import subway.domain.Line;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.station.Station;
 
 import java.util.List;
 import java.util.stream.Collectors;
