+++ src/test/java/subway/domain/StationEdgesTest.java
@@ -2,6 +2,8 @@ package subway.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.domain.line.edge.StationEdge;
+import subway.domain.line.edge.StationEdges;
 
 import java.util.List;
 import java.util.Set;
