+++ src/test/java/subway/domain/SubwayTest.java
@@ -2,6 +2,9 @@ package subway.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.domain.line.Line;
+import subway.domain.line.edge.StationEdge;
+import subway.domain.station.Station;
 
 import java.util.List;
 import java.util.Set;
