+++ src/test/java/subway/repository/StationRepositoryTest.java
@@ -12,7 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import subway.dao.StationDao;
-import subway.domain.Station;
+import subway.domain.station.Station;
 @JdbcTest
 class StationRepositoryTest {
