+++ src/test/java/subway/service/LineServiceTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import subway.domain.Line;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.station.Station;
 import subway.exception.DuplicatedLineNameException;
 import subway.repository.LineRepository;
 import subway.repository.StationRepository;
