+++ src/test/java/subway/service/StationServiceTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import subway.domain.Station;
+import subway.domain.station.Station;
 import subway.exception.DuplicatedStationNameException;
 import subway.repository.StationRepository;
